Extract useful internal manager messaging
In order to send messages to the IML agent and between IML manager
processes, extract a few fns that can be reused between crates.
