+++ Cargo.toml
@@ -1,2 +1,2 @@
 [workspace]
-members = ['iml-wire-types',  'iml-agent-comms', 'iml-rabbit', 'iml-manager-env', 'iml-agent', 'iml-services', 'tokio-runtime-shutdown']
+members = ['iml-wire-types',  'iml-agent-comms', 'iml-rabbit', 'iml-manager-env', 'iml-agent', 'iml-services', 'tokio-runtime-shutdown', 'iml-manager-messaging']
