+++ iml-manager-messaging/Cargo.toml
@@ -1,22 +1,14 @@
 [package]
-name = "iml-agent-comms"
+name = "iml-manager-messaging"
 version = "0.1.0"
 authors = ["IML Team <iml@whamcloud.com>"]
 edition = "2018"
 
 [dependencies]
-env_logger = "0.6.1"
 failure = "0.1.5"
 futures = "0.1.25"
 iml-wire-types = { path = "../iml-wire-types", version = "0.1.0" }
 iml-rabbit = { path = "../iml-rabbit", version = "0.1.0" }
 iml-manager-env = { path = "../iml-manager-env", version = "0.1.0" }
-iml-manager-messaging = { path = "../iml-manager-messaging", version = "0.1.0" }
-lapin-futures = "0.18.0"
-log = "0.4.6"
 serde = { version = "1", features = ["derive"] }
-serde_json = "1.0.39"
-tokio = "0.1.18"
-uuid = { version = "0.7", features = ["v4"] }
-warp = "0.1.14"
-
+serde_json = "1.0"
